requirements/base.txt
https://www.djangoproject.com/download/1.7c2/tarball/
Jinja2==2.7.2
MarkupSafe==0.23
Pillow==2.4.0
Pygments==1.6
Sphinx==1.2.2
beautifulsoup4==4.3.2
dj-database-url==0.3.0
django-appconf==0.6
django-braces==1.4.0
django-configurations==0.8
django-crispy-forms==1.4.0
django-geoposition==0.2.1
django-grappelli==2.5.3
django-leaflet==0.13.5
django-uuidfield==0.5.0
docutils==0.11
easy-thumbnails==2.0.1
envdir==0.6.1
gevent==1.0.1
greenlet==0.4.2
psycopg2==2.5.3
pytz==2014.3
requests==2.2.1
six==1.6.1
wheel==0.23.0
